Assign visual representation of newline \n: ♪ U+266a eighth note
This is not coincidentally codepoint 13 in CP437, a good glyph as any to
visually represent newline, when it is not interpreted as a control
character. From https://en.wikipedia.org/wiki/Eighth_note

"In unicode, the symbols U+266A (♪) and U+266B (♫) are an eighth note and
beamed pair of eighth notes respectively. The former is inherited from
the early 1980s code page 437, where it has code 13."